pyproject.toml
[project]
name = "ej-logiciel-ecommerce"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "celery>=5.5.2",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "reportlab>=4.4.1",
]
